Removed DiscordPreset file, migrated to VideoCompressorService, begin creating tests.
I removed the DiscordPreset. Discord has changed filesize limits and a full-featured compressor will stil do that job. It doesn't make sense to keep working on that and then backpedal. My current testing involves a .ps1 and a file I have locally, so it will be replaced with actual tests soon.